requirements.txt
gradio==4.44.0
requests>=2.31.0
Pillow>=10.0.0
huggingface_hub>=0.30.0
